+++ src/Pact/Types/Gas.hs
@@ -45,6 +45,7 @@ import Pact.Types.Pretty
 import Pact.Types.PactValue
 import Pact.Types.RowData
 import Pact.Types.Term
+import Pact.Types.Namespace
 import Pact.Parse
 import Pact.Types.SizeOf(Bytes)
 
@@ -143,6 +144,10 @@ data GasArgs
   -- ^ Cost of the fold-db call
   | GModuleMemory Bytes
   -- ^ The cost of the in-memory representation of the module
+  | GPrincipal !Int
+  -- ^ the cost of principal creation and validation
+  | GIntegerOpCost !Integer Integer
+  -- ^ Integer costs
 
 instance Pretty GasArgs where
   pretty g = case g of
@@ -162,6 +167,8 @@ instance Pretty GasArgs where
     GDistinct i -> "GDistinct:" <> pretty i
     GFoldDB -> "GFoldDB"
     GModuleMemory i -> "GModuleMemory: " <> pretty i
+    GPrincipal i -> "GPrincipal: " <> pretty i
+    GIntegerOpCost i j -> "GIntegerOpCost:" <> pretty i <> colon <> pretty j
 
 newtype GasLimit = GasLimit ParsedInteger
   deriving (Eq,Ord,Num,Real,Integral,Enum,Serialize,NFData,Generic,ToTerm,ToJSON,Pretty)